+++ integration_tests/tests/integration_tests/main.rs
@@ -3,16 +3,11 @@ mod account_update_tests;
 mod cnft_tests;
 mod cnft_tests_v2_leaf_schema;
 mod common;
+mod das_options_tests;
 mod fungibles_and_token_extensions_tests;
 mod general_scenario_tests;
 mod mpl_core_tests;
-mod nft_editions_tests;
-mod ops_purge;
+mod ops_purge_tests;
 mod regular_nft_tests;
-mod show_collection_metadata_option_tests;
-mod show_fungible_flag_tests;
-mod show_inscription_flag_tests;
-mod test_get_assets_with_multiple_same_ids;
-mod test_show_zero_balance_filter;
 mod token_accounts_tests;
 mod token_type_test;
